Add the ability to not create a Repo with Database
Sometimes when someone needs to create a project just to experiment new
stuff, or access remote endpoints it's not needed a database, and use a
template that creates all files and integrations with a database it's
over engineering. In this commit it's add a new option on checkbox to
the question of "Wwich type of database do you want to use?" that is
"none".