Add rest template
Why this change is needed?
When someone need to prototype a project, sometimes it needs a rest
endpoint to perform some actions. This commit intent to create a
template for the developer to use.

How this commit addess this issue?
It create a new generator inside macchiato. to run yo macchiato:rest.