Add all jars but exclude jar from gradle wrapper #25
Why this change is needed?
When we create a new project that it's using java, we need to exclude
all jars, except the gralde wrapper jar because it's used to run the
build in other machines.

How this commit addess this issue?
It add inside of the .gitignore file a regular expression for all jars
and another one to exclude the gradle wrapper jar.