Create GitHub repository
Why is this change needed?
As part of the production enablement process using GitHub, Travis
and Heroku, the user has to manually create a remote repository on
GitHub and push the application code.

How does this commit address that?
This commit adds the capability of automatically creating the remote
repository on GitHub as part of the cloud sub-generator.