Add default profile to be equals to local (#34)
Why this change is needed?
it someone will run the project inside a IDE, and don't want to specify
inside of the IDE which profile the developer wants too use it, we are
going to assume that the developer wants to use local profile, but for that
to work, if the developer do not specify the profile spring will set as
default, as the project it was already set to only work with local or
production it will break when it's starts. With this commit if the developer does not set a profile it will use as default.

How this commit address this issue?
It set default profile as local when none profile was set.