Remove whitelabel from project
Why this change is needed?
When a error occur the default behavior from spring is to display a
White Label page and as this is for a backend server the most of the
times when someone request from a browser we want to see the json
displayed.

How this commit address this issue?
It disable whitelabel on application,properties