Add GlobalExceptionHandler to provide generic Handlers #29 (#36)
Why this change is needed?
With this commit, we are introducing a global exception handler that
will give for us a handler for spring exceptions. Instead of always sent
a 500 error, with this it will address the correct handler for each type
of exception.

How this commit address this issue?
It adds a ControllerAdvice in the application extendig a generic
exception handler provided by Spring.