Remove files and code if db is not select weren't used #32
Why this change is needed?
it was a bug when the developer selects to create a project without db,
macchiato was generating files and properties that would only be used if
the application was using db.

How this commit address this issue?
It removes all files and properties that was only used by db.